+++ packages/base/Grid/package.json
@@ -35,6 +35,7 @@
     "@material-ui/core": "4.12.4",
     "@toptal/picasso-provider": "*",
     "@toptal/picasso-shared": "14.0.1",
+    "@toptal/picasso-tailwind": ">2.1.0",
     "react": ">=16.12.0 < 19.0.0"
   },
   "exports": {
