+++ packages/base/Grid/src/GridItem/GridItem.tsx
@@ -2,10 +2,10 @@ import type { ReactNode, HTMLAttributes } from 'react'
 import React, { forwardRef, useContext } from 'react'
 import type { GridSize } from '@material-ui/core'
 import { rem, type BaseProps } from '@toptal/picasso-shared'
+import { twMerge } from 'tailwind-merge'
 
 import GridContext from '../GridContext/GridContext'
 import type { GridSpacing } from '../types'
-import { breakpointsList } from '@toptal/picasso-provider'
 
 export interface GridSizes {
   /** Defines the number of grids the component is going to use. It's applied for all the screen sizes with the lowest priority */
@@ -32,13 +32,13 @@ export interface Props
 // TODO: different styles depending on the direction and so on
 // p-0.25 p-0.5 p-0.75 p-1 p-1.25 p-1.5 p-1.75 p-2 p-2.25 p-2.5 p-2.75 p-3 p-3.25 p-3.5 p-3.75 p-4 p-4.25 p-4.5 p-4.75 p-5 p-5.25 p-5.5 p-5.75 p-6 p-6.25 p-6.5 p-6.75 p-7 p-7.25 p-7.5 p-7.75 p-8 p-8.25 p-8.5 p-8.75 p-9 p-9.25 p-9.5 p-9.75 p-10 p-10.25 p-10.5 p-10.75 p-11 p-11.25 p-11.5 p-11.75 p-12 p-12.25 p-12.5 p-12.75 p-13 p-13.25 p-13.5 p-13.75 p-14 p-14.25 p-14.5 p-14.75 p-15 p-15.25 p-15.5 p-15.75 p-16 p-16.25 p-16.5 p-16.75 p-17 p-17.25 p-17.5 p-17.75 p-18 p-18.25 p-18.5 p-18.75 p-19 p-19.25 p-19.5 p-19.75 p-20 p-20.25 p-20.5 p-20.75 p-21 p-21.25 p-21.5 p-21.75 p-22 p-22.25 p-22.5 p-22.75 p-23 p-23.25 p-23.5 p-23.75 p-24
 const getSpacingClassName = (spacing: GridSpacing) => {
-
   // TODO: recheck, if it is so simple as it looks like
   console.log('@@@ spacing', spacing)
   if (spacing === 0) {
     return ''
   }
 
+  // TODO: recheck, px / rem problem
   return `p-${rem(spacing * 2).replace('rem', '')}`
 }
 
@@ -55,41 +55,44 @@ const getSpacingClassName = (spacing: GridSpacing) => {
 // lg:max-w-1/12 lg:max-w-2/12 lg:max-w-3/12 lg:max-w-4/12 lg:max-w-5/12 lg:max-w-6/12 lg:max-w-7/12 lg:max-w-8/12 lg:max-w-9/12 lg:max-w-10/12 lg:max-w-11/12 lg:max-w-full
 // xl:max-w-1/12 xl:max-w-2/12 xl:max-w-3/12 xl:max-w-4/12 xl:max-w-5/12 xl:max-w-6/12 xl:max-w-7/12 xl:max-w-8/12 xl:max-w-9/12 xl:max-w-10/12 xl:max-w-11/12 xl:max-w-full
 
-const generateClassesForNonXsBreakpoints = (breakpointKey: string, breakpointValue?: GridSize | boolean) => {
+const generateClassesForNonXsBreakpoints = (
+  breakpointKey: string,
+  breakpointValue?: GridSize | boolean
+) => {
   if (breakpointValue === true) {
-    throw new Error('TBD')
-    // For the auto layouting
+    throw new Error('TODO')
+    // For the auto layout
     // styles = {
     //   flexBasis: 0,
     //   flexGrow: 1,
     //   maxWidth: '100%',
     // };
   } else if (breakpointValue === 'auto') {
-    throw new Error('TBD')
+    throw new Error('TODO')
   } else if (breakpointValue === undefined) {
     return []
   }
 
   console.log('@@@ breakpointValue', breakpointValue, typeof breakpointValue)
 
-  const breakpointValueClassName = breakpointValue === 12 ? 'full' : `${breakpointValue}/12`
+  const breakpointValueClassName =
+    breakpointValue === 12 ? 'full' : `${breakpointValue}/12`
 
-  return [`${breakpointKey}:basis-${breakpointValueClassName} ${breakpointKey}:max-w-${breakpointValueClassName}`]
+  return [
+    `${breakpointKey}:basis-${breakpointValueClassName} ${breakpointKey}:max-w-${breakpointValueClassName}`,
+  ]
 }
 
 // TODO: rework, going with ifs for now
 // TODO: different styles depending on the direction and so on
 // Generates styling for one particular item based on the current breakpoint and the number of columns it should take
 const getBreakpointAndColumnClassName = ({ xs, sm, md, lg, xl }: GridSizes) => {
-
-
   // TODO: consider boolean and 'auto' values
-
   let classNames = ['grow-0']
-  console.log('@@@ breakpointsList', breakpointsList)
+
   if (xs) {
     if (xs === true) {
-      throw new Error('TBD')
+      throw new Error('TODO')
       // For the auto layouting
       // styles = {
       //   flexBasis: 0,
@@ -97,16 +100,13 @@ const getBreakpointAndColumnClassName = ({ xs, sm, md, lg, xl }: GridSizes) => {
       //   maxWidth: '100%',
       // };
     } else if (xs === 'auto') {
-      throw new Error('TBD')
+      throw new Error('TODO')
     }
-
-
-    console.log('@@@ xs', xs, typeof xs)
 
     const xsClassName = xs === 12 ? 'full' : `${xs}/12`
 
     // Assuming that xs is a number from now on
-    classNames.push(`basis-${xsClassName} max-w-${xsClassName}`) 
+    classNames.push(`basis-${xsClassName} max-w-${xsClassName}`)
   }
 
   const smClasses = generateClassesForNonXsBreakpoints('sm', sm)
@@ -114,9 +114,8 @@ const getBreakpointAndColumnClassName = ({ xs, sm, md, lg, xl }: GridSizes) => {
   const lgClasses = generateClassesForNonXsBreakpoints('lg', lg)
   const xlClasses = generateClassesForNonXsBreakpoints('xl', xl)
 
-  classNames = classNames.concat(smClasses, mdClasses, lgClasses, xlClasses) 
+  classNames = classNames.concat(smClasses, mdClasses, lgClasses, xlClasses)
 
-  console.log('@@@ classNames', classNames)
   return classNames
 }
 
@@ -129,33 +128,30 @@ export const GridItem = forwardRef<HTMLDivElement, Props>(function GridItem(
   const { spacing } = useContext(GridContext)
   const spacingClassName = getSpacingClassName(spacing as GridSpacing)
 
-  const breakpointAndColumnClassName = getBreakpointAndColumnClassName({xs, sm, md, lg, xl}).join(' ')
+  const breakpointAndColumnClassName = getBreakpointAndColumnClassName({
+    xs,
+    sm,
+    md,
+    lg,
+    xl,
+  }).join(' ')
 
+  return (
+    <div
+      ref={ref}
+      className={twMerge(
+        spacingClassName,
+        breakpointAndColumnClassName,
+        className
+      )}
+      style={style}
+      {...rest}
+    >
+      {children}
+    </div>
+  )
 
-  return (<div
-    className={`${spacingClassName} ${breakpointAndColumnClassName} `}
-    style={style}
-  >
-    {children}
-  </div>)
-
-  // return (
-  //   <MUIGrid
-  //     {...rest}
-  //     ref={ref}
-  //     item
-  //     xs={xs}
-  //     sm={sm}
-  //     md={md}
-  //     lg={lg}
-  //     xl={xl}
-  //     classes={classes}
-  //     className={className}
-  //     style={style}
-  //   >
-  //     {children}
-  //   </MUIGrid>
-  // )
+  // return (<MUIGrid item ...>{children}</MUIGrid>)
 })
 
 GridItem.defaultProps = {}
