+++ packages/base/Grid/src/types.ts
@@ -1,1 +1,25 @@
-export type GridSpacing = 0 | 8 | 16 | 32 | 64 | 72 | 80;
+// TODO: where is this type coming from?
+export type GridSpacing = 0 | 8 | 16 | 32 | 64 | 72 | 80
+
+// Taken from MUI@4 to maintain API compatibility
+export type GridJustification =
+  | 'flex-start'
+  | 'center'
+  | 'flex-end'
+  | 'space-between'
+  | 'space-around'
+  | 'space-evenly'
+
+// Taken from MUI@4 to maintain API compatibility
+export type GridItemsAlignment =
+  | 'flex-start'
+  | 'center'
+  | 'flex-end'
+  | 'stretch'
+  | 'baseline'
+
+// Taken from MUI@4 to maintain API compatibility
+export type GridDirection = 'row' | 'row-reverse' | 'column' | 'column-reverse'
+
+// Taken from MUI@4 to maintain API compatibility
+export type GridWrap = 'nowrap' | 'wrap' | 'wrap-reverse'
