+++ packages/base/Grid/tsconfig.json
@@ -4,6 +4,7 @@
   "include": ["src"],
   "references": [
     { "path": "../../picasso-provider" },
+    { "path": "../../picasso-tailwind" },
     { "path": "../../shared" },
     { "path": "../Icons" },
     { "path": "../Typography" },
